Appsheet Function

This function interacts with the AppSheet API to perform actions on a specified table. For more info, check the official documentation.

appsheet(
  tableName,
  Action = "Find",
  Properties = ash_properties(),
  Rows = list(),
  appId = Sys.getenv("APPSHEET_APP_ID"),
  access_key = Sys.getenv("APPSHEET_APP_ACCESS_KEY")
)

Arguments

tableName

The name of the table to perform actions on.

Action

The action to be performed on the table, one of ("Find", "Add", "Delete", "Edit"). Default is "Find", which reads a table.

Properties

A list of properties for the action. ash_properties() provides sensible defaults, but can be customized.

Rows

A list of rows for the action. Default is an empty list.

appId

The AppSheet application ID. Default is retrieved from the APPSHEET_APP_ID environment variable.

access_key

The AppSheet application access key. Default is retrieved from the APPSHEET_APP_ACCESS_KEY environment variable.

Value

A data frame containing the response from the AppSheet API.

Examples

if (FALSE) {
appsheet("my_table")
appsheet("my_table", Properties = ash_properties(Locale = "en-GB"))
}