Build the properties of an Appsheet request

This function exist to provide auto completion. Still, it is likely that users will need to visit the official documentation for further customization.

ash_properties(
  Locale = "en-US",
  Location = NULL,
  RunAsUserEmail = NULL,
  Timezone = NULL,
  UserSettings = NULL,
  Selector = NULL,
  ...
)

Arguments

Locale

Locale of the client making the request. For example, en-US indicates English, United States. If this value is not specified, Locale defaults to en-US.
The Locale is used when validating Date, DateTime, Decimal, Percent, Price, and Time data values. For example, when Locale is en-US, date values must be entered in MM/DD/YYYY format; when Locale is en-GB, date values must be entered in DD/MM/YYYY format.

Location

Latitude and longitude of the client making the request. If not specified, Location defaults to 0.000000, 0.000000.

RunAsUserEmail

User email address of the person taking the action. The action is performed under the user email address you specify. If this value is not specified, the action is performed under the user email address of the application owner.

Timezone

Timezone of the client making the request.

UserSettings

User setting to be used when performing the action. If not specified, no user settings are used.

Selector

Expression to select and format the rows returned. Only valid when Action is "Find".

...

Placeholder for future properties to be introduced in the API.

Value

A list of properties

Details

Any NULL will be internally dropped before performing the request.

Examples

ash_properties()
#> $Locale
#> [1] "en-US"
#> 
ash_properties(Locale = "en-GB")
#> $Locale
#> [1] "en-GB"
#>